feat(user): register user in centralised table with extra information
Due to the infrastructure of the homer SMS system, users are stored in
different tables for each agency/ministry enforcing their
SHN compliance.

Using the native app would require a centralised source of user info currently registered to the native app. Hence I thought it'd be better to
store the user's contact number in a new table, along with the agency
that is enforcing their SHN.

Along with that, the user's push notification token is also stored for
push notification location pinging.