feat(user): use constant that stores table names when finding a user
Previously when finding a user, we had an array of agency prefixes
(`ica-homer`, `homer`, `mom-dloa`, etc.), where we'd loop through them
and appended `-users-<table_env>` to it before making a query to find
the user.

This commit now introduces an array with all the users table already
constructed (`ALL_HOMER_USERS_TABLE`). Now, when finding a user, we do
not need to construct the table from the agency prefix and can just loop
through that above-mentioned constant instead.