fix: update OtpService.saveOtp
This fix updates the OtpService's saveOtp to accept an object instead
of multiple params. This change accomodates an additional email attribute
which is passed if the OTP is being generated for an admin user.

With the new changes, the contactNumber attribute will be empty if the
OTP is being generated for an admin user, and the email attribute will be
empty if the OTP is being generated for a user under quarantine orders