feat: update OTP verification to accept both emails and phone numbers
This commit updates the query used in the OtpService method getOtp to
query correctly for an email or a phone number based on an `isAdmin`
argument.

It also updates the OTP verification route to verify OTP correctly for
both emails and phone numbers.

Finally, it modifies the parameters set for the access token depending
on whether the OTP is for an admin user or someone under quarantine. The
admin user's access token contains a `role` attribute with the value
`admin` as opposed to the `order_id` attribute for someone under quarantine.