feat: update auth middleware to differentiate between admin and QO users
This commit wraps the auth middleware in a function which accepts an
argument `isAdmin`. If `isAdmin` is `true`, then the returned middleware
function checks for whether the access token contains an attribute `role`
with the value `admin`. If it does not, it'll throw an error and send an
error response.