Feat/add admin otp flow (#10)
* feat: add model for admin users

* fix: make email unique for AdminUsers model

* feat: AdminUserService

* fix: update OtpService.saveOtp

This fix updates the OtpService's saveOtp to accept an object instead
of multiple params. This change accomodates an additional email attribute
which is passed if the OTP is being generated for an admin user.

With the new changes, the contactNumber attribute will be empty if the
OTP is being generated for an admin user, and the email attribute will be
empty if the OTP is being generated for a user under quarantine orders

* fix: update OTP model and add migration

This commit updates the OTP model to include a nullable field, `email`, as well
as change the existing `contact_number` field to be nullable. The commit also
includes the necessary migrations for the aforementioned changes.

* feat: update OTP route to accomodate admin users

* chore: fix typos in migration files

* chore: add seed for AdminUsers

* feat: update OTP verification to accept both emails and phone numbers

This commit updates the query used in the OtpService method getOtp to
query correctly for an email or a phone number based on an `isAdmin`
argument.

It also updates the OTP verification route to verify OTP correctly for
both emails and phone numbers.

Finally, it modifies the parameters set for the access token depending
on whether the OTP is for an admin user or someone under quarantine. The
admin user's access token contains a `role` attribute with the value
`admin` as opposed to the `order_id` attribute for someone under quarantine.

* feat: update auth middleware to differentiate between admin and QO users

This commit wraps the auth middleware in a function which accepts an
argument `isAdmin`. If `isAdmin` is `true`, then the returned middleware
function checks for whether the access token contains an attribute `role`
with the value `admin`. If it does not, it'll throw an error and send an
error response.

* feat: create and implement custom error for auth middleware

* chore: seed db data with later end dates for ease of testing logic

* chore: correct adminUsers migration

* chore: update API documentation

* fix: requested changes on PR

* feat: add request validation for otp routes

* fix: typo

* fix: otp validation schema