Feat/add otp sms functionality (#11)
* chore: install dependencies

* feat: update OTP verification to accept both emails and phone numbers

This commit updates the query used in the OtpService method getOtp to
query correctly for an email or a phone number based on an `isAdmin`
argument.

It also updates the OTP verification route to verify OTP correctly for
both emails and phone numbers.

Finally, it modifies the parameters set for the access token depending
on whether the OTP is for an admin user or someone under quarantine. The
admin user's access token contains a `role` attribute with the value
`admin` as opposed to the `order_id` attribute for someone under quarantine.

* feat: update auth middleware to differentiate between admin and QO users

This commit wraps the auth middleware in a function which accepts an
argument `isAdmin`. If `isAdmin` is `true`, then the returned middleware
function checks for whether the access token contains an attribute `role`
with the value `admin`. If it does not, it'll throw an error and send an
error response.

* chore: update env example

* feat: add SMS functionality

* feat: add request validation for otp routes

* fix: otp validation schema

* fix: add alpha sender ID

* fix: update package-lock